requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
cffi
cryptography
Django==2.1.2
django-redis==4.9.0
djangorestframework==3.8.2
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient
pycparser==2.19
pylint==2.1.1
PyMySQL==0.9.2
pytz==2018.5
redis==2.10.6
six==1.11.0
typed-ast
wrapt==1.10.11
